refactor to internal API, separate simulated time tests into different Specs
